Add an inverse implementation of updateState with property test.
One way we could implement time travel is by maintaining an inverse
implementation of updateState.

Think about it like this:

const initialState = <whatever>;

const updateArguments = <whatever>;
const updatedState = updateState(initialState, updateArguments);

const reversedState = reverseState(updatedState, updateArguments);

reversedState should be the same as initialState.

The property test ensures this is true across all possible sequences of
input.

This is a tricky thing to do though, as no state update can lose any
information that is needed to reverse the process.

It remains to be seen if any complex behaviour can work with this
approach but it's fun to investigate.